+++ versions.tf
@@ -18,15 +18,11 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 5.75"
+      version = "5.75.0"
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"
-      version = ">= 2.32"
-    }
-    helm = {
-      source  = "hashicorp/helm"
-      version = ">= 2.16"
+      version = "2.32.0"
     }
   }
 }